+++ src/main/java/school/faang/BJS2_79150/User.java
@@ -1,0 +1,38 @@
+package school.faang.BJS2_79150;
+
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+
+import lombok.Getter;
+import lombok.AllArgsConstructor;
+import lombok.ToString;
+
+@ToString
+@Getter
+@AllArgsConstructor
+public class User {
+
+    private String name;
+    private int age;
+    private String workplace;
+    private String address;
+
+
+
+    public static Map<Integer, List<User>> groupUsers(List<User> users) {
+        Map<Integer, List<User>> result = new HashMap<>();
+
+        for (User user : users) {
+            int age = user.age;
+
+            if (!result.containsKey(age)) {
+                result.put(age, new ArrayList<>());
+            }
+            result.get(age).add(user);
+        }
+
+        return result;
+    }
+}
